Move hardware sine waves to examples
