+++ resources/src/ICRA2015/learning/edgeConfig.ini
