+++ resources/src/ICRA2015/learning/logs/.gitignore
