+++ resources/src/ICRA2015/learning/nodeConfig.ini
