+++ resources/src/tetraTerrain/controlVars.json
@@ -1,3 +1,4 @@
+
 {
 	"body_waves" : 1,
 	"cpg_frequency" : 0.5,
@@ -5,12 +6,13 @@
 	"in_left_offset" : 0,
 	"in_right_offset" : 0,
 	"in_top_amp_a" : -30,
-	"in_top_offset" : 1.5700000000000001,
+	"in_top_offset" : 1.57,
 	"out_bottom_amp_a" : 100,
 	"out_left_offset" : 0,
 	"out_right_offset" : 0,
 	"out_top_amp_a" : 100,
-	"out_top_offset" : 1.5700000000000001,
+	"out_top_offset" : 1.57,
 	"scores" : 0,
 	"updateFrequency" : 0.01
 }
+
