+++ src/examples/motorModel/tsTestRig.cpp
