+++ test_integration/TimestepIndependence/MotorTimestep_test.cpp
@@ -27,7 +27,7 @@
 */
 
 // This application
-#include "dev/btietz/timestepTest/tsTestRig.h"
+#include "examples/motorModel/tsTestRig.h"
 // This library
 #include "core/tgSpringCableActuator.h"
 #include "core/tgModel.h"
